refactor(desktop): extract Transcript tail components / 拆分 Transcript 尾部组件
The live-turn footer work pushed Transcript.tsx to 1247 lines, 167 over
its repolint file-size budget (1080). Extract the self-contained tail
components into their own modules — ProcessFoldHeader, QuestionJumpBar,
and the row cards (PhaseCard/SteerCard/NoticeCard/CompactionCard) —
bringing the file to 945 lines. NoticeCard keeps its historical export
path from Transcript.tsx. No behavior change.

Verification: repolint clean; tsc + eslint clean; pnpm test passes all
175 suites.
