+++ desktop/frontend/src/components/QuestionJumpBar.tsx
@@ -1,0 +1,133 @@
+// QuestionJumpBar: the question navigator rail along the transcript edge.
+
+import { useEffect, useRef, useState, type CSSProperties, type MouseEvent as ReactMouseEvent } from "react";
+import { useT } from "../lib/i18n";
+import type { QuestionAnchor } from "../lib/transcriptGrouping";
+
+export function QuestionJumpBar({ questions, onJump }: { questions: QuestionAnchor[]; onJump: (question: QuestionAnchor) => void }) {
+  const t = useT();
+  const [hovered, setHovered] = useState<number | null>(null);
+  const [active, setActive] = useState<number | null>(null);
+  const barRef = useRef<HTMLDivElement>(null);
+  const previewTop = useRef(0);
+  const [showPreview, setShowPreview] = useState(false);
+
+  useEffect(() => {
+    if (questions.length === 0) return;
+    setActive(questions[questions.length - 1]?.turn ?? null);
+  }, [questions]);
+
+  useEffect(() => {
+    if (active === null) return;
+    const el = barRef.current?.querySelector(`[data-turn="${active}"]`);
+    el?.scrollIntoView({ block: "nearest" });
+  }, [active]);
+
+  const hoverIdx = hovered !== null ? questions.findIndex((question) => question.turn === hovered) : -1;
+  const hoveredQuestion = hovered !== null ? questions.find((question) => question.turn === hovered) : undefined;
+
+  const closestQuestionFromY = (clientY: number): { question: QuestionAnchor; previewY: number } | null => {
+    const el = barRef.current;
+    if (!el) return null;
+    const markers = el.querySelectorAll<HTMLElement>(".jump-item");
+    const barRect = el.getBoundingClientRect();
+    let closest = -1;
+    let closestDist = Infinity;
+    let closestY = 0;
+    markers.forEach((item, index) => {
+      const rect = item.getBoundingClientRect();
+      const midY = rect.top + rect.height / 2;
+      const dist = Math.abs(clientY - midY);
+      if (dist < closestDist) {
+        closestDist = dist;
+        closest = index;
+        closestY = midY - barRect.top;
+      }
+    });
+    const question = questions[closest];
+    if (!question) return null;
+    return { question, previewY: closestY };
+  };
+
+  const onMove = (e: ReactMouseEvent<HTMLDivElement>) => {
+    const closest = closestQuestionFromY(e.clientY);
+    if (!closest) return;
+    previewTop.current = closest.previewY;
+    setHovered(closest.question.turn);
+    setShowPreview(true);
+  };
+
+  const scrollTo = (question: QuestionAnchor) => {
+    setActive(question.turn);
+    onJump(question);
+  };
+
+  const onRailMouseDown = (e: ReactMouseEvent<HTMLDivElement>) => {
+    const closest = closestQuestionFromY(e.clientY);
+    if (!closest) return;
+    e.preventDefault();
+    previewTop.current = closest.previewY;
+    setHovered(closest.question.turn);
+    setShowPreview(true);
+    scrollTo(closest.question);
+  };
+
+  const onItemMouseDown = (e: ReactMouseEvent<HTMLButtonElement>, question: QuestionAnchor) => {
+    e.preventDefault();
+    scrollTo(question);
+  };
+
+  const dotProps = (
+    idx: number,
+    turn: number,
+  ): { style: CSSProperties; "data-d"?: string } => {
+    const isActive = active === turn;
+    if (hoverIdx < 0) {
+      return { style: { width: isActive ? 18 : 12, background: isActive ? "var(--accent)" : undefined } };
+    }
+    const d = Math.abs(idx - hoverIdx);
+    const width = d === 0 ? 32 : d === 1 ? 20 : d === 2 ? 14 : isActive ? 18 : 12;
+    const background = d <= 2 ? undefined : isActive ? "var(--accent)" : undefined;
+    return {
+      style: { width, transitionDelay: `${d * 20}ms`, background },
+      "data-d": d <= 2 ? String(d) : undefined,
+    };
+  };
+
+  return (
+    <nav
+      className="jump-bar"
+      ref={barRef}
+      aria-label={t("questionNav.label")}
+      onMouseMove={onMove}
+      onMouseLeave={() => {
+        setHovered(null);
+        setShowPreview(false);
+      }}
+    >
+      <div className="jump-scroll" onMouseDown={onRailMouseDown} onClick={onRailMouseDown}>
+        {questions.map((question, index) => (
+          <button
+            className="jump-item"
+            key={question.id}
+            type="button"
+            data-turn={question.turn}
+            aria-label={t("questionNav.jump", { n: question.turn + 1 })}
+            onMouseDown={(e) => onItemMouseDown(e, question)}
+            onClick={(e) => {
+              e.stopPropagation();
+              if (e.detail === 0) scrollTo(question);
+            }}
+          >
+            <span className="jump-dot" {...dotProps(index, question.turn)} />
+          </button>
+        ))}
+      </div>
+      {showPreview && hoveredQuestion && (
+        <div className="jump-preview" style={{ top: previewTop.current }} role="tooltip">
+          <span className="jump-text">{hoveredQuestion.text}</span>
+        </div>
+      )}
+    </nav>
+  );
+}
