+++ desktop/frontend/src/components/Transcript.tsx
@@ -1,19 +1,18 @@
-import { forwardRef, memo, type CSSProperties, type MouseEvent as ReactMouseEvent, type PointerEvent as ReactPointerEvent, type ReactNode, useCallback, useContext, useEffect, useMemo, useRef, useState, useSyncExternalStore } from "react";
+import { forwardRef, memo, type CSSProperties, type PointerEvent as ReactPointerEvent, type ReactNode, useCallback, useContext, useEffect, useMemo, useRef, useState, useSyncExternalStore } from "react";
 import { Virtuoso, type Components, type ItemProps, type ListItem, type ListProps } from "react-virtuoso";
 import type { ControllerLiveStore, Item, LiveStream } from "../lib/useController";
 import type { CheckpointMeta } from "../lib/types";
 import type { InvocationMetadataMap } from "../lib/invocationDisplay";
 import { useT } from "../lib/i18n";
 import { AssistantMessage, InvocationMetadataContext, TurnActions, UserMessage } from "./Message";
-import { ProcessCompactIcon, ProcessPhaseIcon } from "./ProcessCard";
 import { ToolCard } from "./ToolCard";
 import { ExtensionCard } from "./ExtensionCard";
-import { ArrowDown, ChevronRight, CirclePlay, FileSearch, Info, TriangleAlert } from "lucide-react";
+import { ArrowDown } from "lucide-react";
 import { Welcome } from "./Welcome";
 import { ReadOnlyBatch } from "./ReadOnlyBatch";
 import { ToolGroup } from "./ToolGroup";
 import { getProcessFoldPreference, onProcessFoldPreferenceChange, type ProcessFoldPreference } from "../lib/processFoldPreference";
-import { STEER_NOTICE_PREFIX, isSteerNoticeText } from "../lib/useController";
+import { isSteerNoticeText } from "../lib/useController";
 import { useTranscriptEntranceAnimation } from "../lib/useEntranceAnimation";
 import { useTranscriptSelectionRetention } from "../lib/useTranscriptSelectionRetention";
 import { compactQuestionText, lastQuestionTurn, questionAnchorId, questionTurnsById, type QuestionAnchor } from "../lib/transcriptGrouping";
@@ -32,8 +31,6 @@ import {
   NO_LIVE,
   type AssistantItem,
   type FoldMap,
-  type NoticeItem,
-  type SegmentModel,
   type ToolItem,
   type TranscriptLiveFlags,
   type TranscriptRow,
@@ -44,7 +41,9 @@ import { noteTranscriptRowCounts } from "../lib/sessionDiagnostics";
 import { useReasoningDisplayMode } from "../lib/reasoningDisplayPreference";
 import { InlineAssistantReasoning } from "./InlineAssistantReasoning";
 import { LiveTurnRegion } from "./LiveTurnRegion";
-import { useTick, workStatusLabel } from "../lib/workStatus";
+import { ProcessFoldHeader } from "./ProcessFoldHeader";
+import { QuestionJumpBar } from "./QuestionJumpBar";
+import { CompactionCard, NoticeCard, PhaseCard, SteerCard } from "./TranscriptCards";
 import { LiveStreamContext } from "./LiveStreamContext";
 import { useTranscriptSelectableRows } from "../lib/useTranscriptSelectableRows";
 import { TranscriptSelectionOverlay } from "./TranscriptSelectionOverlay";
@@ -52,8 +51,11 @@ import { useCreationTranscriptScrollbar } from "../lib/useCreationTranscriptScro
 import { useTranscriptScrollInteractions } from "../lib/useTranscriptScrollInteractions";
 import { hasTranscriptScrollableRange, TRANSCRIPT_AT_BOTTOM_THRESHOLD_PX, useTranscriptVirtuosoScroll } from "../lib/useTranscriptVirtuosoScroll";
 import { useTranscriptVirtuosoRecovery } from "../lib/useTranscriptVirtuosoRecovery";
-import { TranscriptLayoutIntentProvider, useTranscriptUserResizeIntent } from "./TranscriptLayoutIntentContext";
+import { TranscriptLayoutIntentProvider } from "./TranscriptLayoutIntentContext";
 import { MarkdownImageTabContext } from "./MarkdownImageContext";
+
+// NoticeCard lives with the other row cards; keep the historical export path.
+export { NoticeCard } from "./TranscriptCards";
 type OpenTurnAction = { turn: number; menu: "summary" | "rewind" };
 const QUESTION_NAV_MIN_COUNT = 2;
 type AssistantReasoningDisplay = "normal" | "hide";
@@ -942,306 +944,5 @@ export function Transcript({
     </TranscriptLayoutIntentProvider>
     </MarkdownImageTabContext.Provider>
     </InvocationMetadataContext.Provider>
-  );
-}
-
-// ── ProcessFoldHeader: the fold header row of one process segment ────────────
-// The fold body is NOT rendered here: an open fold contributes its body rows
-// to the virtual row model (they mount only when scrolled into view), a closed
-// fold builds no React subtree at all.
-
-function ProcessFoldHeader({
-  segment,
-  open,
-  onToggle,
-  turnStartAt,
-}: {
-  segment: SegmentModel;
-  open: boolean;
-  onToggle: () => void;
-  turnStartAt?: number;
-}) {
-  const t = useT();
-  const live = useContext(LiveStreamContext);
-  const displayItems = segment.displayItems;
-
-  const hasRunningWork = segment.hasRunningWork;
-  const now = useTick(hasRunningWork);
-  const runningDurationMs = hasRunningWork
-    ? turnStartAt
-      ? Math.max(0, now - turnStartAt)
-      : live?.reasoningStartedAt
-        ? Math.max(0, now - live.reasoningStartedAt)
-        : 0
-    : 0;
-  const effectiveDurationMs = hasRunningWork ? Math.max(segment.durationMs, runningDurationMs) : segment.durationMs;
-
-  const baseLabel = workStatusLabel(effectiveDurationMs, hasRunningWork, t);
-  // Surface what the closed fold hides — a bare duration reads as pure timing
-  // and users have no way to know process detail sits behind it.
-  const toolCount = displayItems.reduce((n, it) => n + (it.kind === "tool" ? 1 : 0), 0);
-  const thoughtCount = displayItems.reduce((n, it) => n + (it.kind === "assistant" ? 1 : 0), 0);
-  const countParts: string[] = [];
-  if (toolCount > 0) countParts.push(t("transcript.toolCount", { n: toolCount }));
-  if (thoughtCount > 0) countParts.push(t("transcript.thoughtCount", { n: thoughtCount }));
-  const label = segment.labelStyle === "counts"
-    ? (countParts.length > 0 ? countParts.join(" · ") : t("transcript.processed"))
-    : countParts.length > 0
-      ? `${baseLabel} · ${countParts.join(" · ")}`
-      : baseLabel;
-  return (
-    <div className={`turn-collapse${open ? " turn-collapse--open" : ""}`} data-kind="reasoning" data-entrance={displayItems[0]?.id || undefined}>
-      <button
-        type="button"
-        className="reasoning__head"
-        onClick={onToggle}
-        aria-expanded={open}
-      >
-        <span className="turn-collapse__label" data-creation-label={label}>{label}</span>
-        {!hasRunningWork && <ChevronRight className={`reasoning__chevron${open ? " reasoning__chevron--open" : ""}`} size={12} />}
-      </button>
-    </div>
-  );
-}
-
-// ── JumpBar, PhaseCard, NoticeCard, CompactionCard ────────────────────────────
-
-function QuestionJumpBar({ questions, onJump }: { questions: QuestionAnchor[]; onJump: (question: QuestionAnchor) => void }) {
-  const t = useT();
-  const [hovered, setHovered] = useState<number | null>(null);
-  const [active, setActive] = useState<number | null>(null);
-  const barRef = useRef<HTMLDivElement>(null);
-  const previewTop = useRef(0);
-  const [showPreview, setShowPreview] = useState(false);
-
-  useEffect(() => {
-    if (questions.length === 0) return;
-    setActive(questions[questions.length - 1]?.turn ?? null);
-  }, [questions]);
-
-  useEffect(() => {
-    if (active === null) return;
-    const el = barRef.current?.querySelector(`[data-turn="${active}"]`);
-    el?.scrollIntoView({ block: "nearest" });
-  }, [active]);
-
-  const hoverIdx = hovered !== null ? questions.findIndex((question) => question.turn === hovered) : -1;
-  const hoveredQuestion = hovered !== null ? questions.find((question) => question.turn === hovered) : undefined;
-
-  const closestQuestionFromY = (clientY: number): { question: QuestionAnchor; previewY: number } | null => {
-    const el = barRef.current;
-    if (!el) return null;
-    const markers = el.querySelectorAll<HTMLElement>(".jump-item");
-    const barRect = el.getBoundingClientRect();
-    let closest = -1;
-    let closestDist = Infinity;
-    let closestY = 0;
-    markers.forEach((item, index) => {
-      const rect = item.getBoundingClientRect();
-      const midY = rect.top + rect.height / 2;
-      const dist = Math.abs(clientY - midY);
-      if (dist < closestDist) {
-        closestDist = dist;
-        closest = index;
-        closestY = midY - barRect.top;
-      }
-    });
-    const question = questions[closest];
-    if (!question) return null;
-    return { question, previewY: closestY };
-  };
-
-  const onMove = (e: ReactMouseEvent<HTMLDivElement>) => {
-    const closest = closestQuestionFromY(e.clientY);
-    if (!closest) return;
-    previewTop.current = closest.previewY;
-    setHovered(closest.question.turn);
-    setShowPreview(true);
-  };
-
-  const scrollTo = (question: QuestionAnchor) => {
-    setActive(question.turn);
-    onJump(question);
-  };
-
-  const onRailMouseDown = (e: ReactMouseEvent<HTMLDivElement>) => {
-    const closest = closestQuestionFromY(e.clientY);
-    if (!closest) return;
-    e.preventDefault();
-    previewTop.current = closest.previewY;
-    setHovered(closest.question.turn);
-    setShowPreview(true);
-    scrollTo(closest.question);
-  };
-
-  const onItemMouseDown = (e: ReactMouseEvent<HTMLButtonElement>, question: QuestionAnchor) => {
-    e.preventDefault();
-    scrollTo(question);
-  };
-
-  const dotProps = (
-    idx: number,
-    turn: number,
-  ): { style: CSSProperties; "data-d"?: string } => {
-    const isActive = active === turn;
-    if (hoverIdx < 0) {
-      return { style: { width: isActive ? 18 : 12, background: isActive ? "var(--accent)" : undefined } };
-    }
-    const d = Math.abs(idx - hoverIdx);
-    const width = d === 0 ? 32 : d === 1 ? 20 : d === 2 ? 14 : isActive ? 18 : 12;
-    const background = d <= 2 ? undefined : isActive ? "var(--accent)" : undefined;
-    return {
-      style: { width, transitionDelay: `${d * 20}ms`, background },
-      "data-d": d <= 2 ? String(d) : undefined,
-    };
-  };
-
-  return (
-    <nav
-      className="jump-bar"
-      ref={barRef}
-      aria-label={t("questionNav.label")}
-      onMouseMove={onMove}
-      onMouseLeave={() => {
-        setHovered(null);
-        setShowPreview(false);
-      }}
-    >
-      <div className="jump-scroll" onMouseDown={onRailMouseDown} onClick={onRailMouseDown}>
-        {questions.map((question, index) => (
-          <button
-            className="jump-item"
-            key={question.id}
-            type="button"
-            data-turn={question.turn}
-            aria-label={t("questionNav.jump", { n: question.turn + 1 })}
-            onMouseDown={(e) => onItemMouseDown(e, question)}
-            onClick={(e) => {
-              e.stopPropagation();
-              if (e.detail === 0) scrollTo(question);
-            }}
-          >
-            <span className="jump-dot" {...dotProps(index, question.turn)} />
-          </button>
-        ))}
-      </div>
-      {showPreview && hoveredQuestion && (
-        <div className="jump-preview" style={{ top: previewTop.current }} role="tooltip">
-          <span className="jump-text">{hoveredQuestion.text}</span>
-        </div>
-      )}
-    </nav>
-  );
-}
-
-type CompactionItem = Extract<Item, { kind: "compaction" }>;
-
-function PhaseCard({ id, text }: { id: string; text: string }) {
-  return <div className="phase" data-entrance={id}><ProcessPhaseIcon size={12} /><span>{text}</span></div>;
-}
-
-// A mid-turn steer is the user's own message, so it renders on the user side
-// of the transcript instead of disappearing into the work fold.
-function SteerCard({ id, text }: { id: string; text: string }) {
-  const t = useT();
-  const body = text.startsWith(STEER_NOTICE_PREFIX) ? text.slice(STEER_NOTICE_PREFIX.length) : text;
-  return (
-    <div className="steer-line" data-entrance={id}>
-      <div className="steer-line__bubble" title={t("transcript.steer")}>
-        <span className="steer-line__icon" aria-hidden="true">↪</span>
-        <span className="steer-line__text">{body}</span>
-      </div>
-    </div>
-  );
-}
-
-function DecisionReceiptLine({ receipt }: { receipt: NonNullable<NoticeItem["decisionReceipt"]> }) {
-  const t = useT();
-  const titleKey = receipt.kind === "ask"
-    ? "notice.decisionReceiptAsk"
-    : receipt.kind === "plan"
-    ? "notice.decisionReceiptPlan"
-    : receipt.kind === "recovery"
-    ? "notice.decisionReceiptRecovery"
-    : "notice.decisionReceiptTool";
-  const outcomeKeys: Record<string, string> = {
-    allow_once: "notice.decisionAllowOnce",
-    allow_session: "notice.decisionAllowSession",
-    allow_persistent: "notice.decisionAllowPersistent",
-    deny: "notice.decisionDeny",
-    start_execution: "notice.decisionStartExecution",
-    revise_plan: "notice.decisionRevisePlan",
-    exit_plan: "notice.decisionExitPlan",
-    recovery_continue: "notice.decisionRecoveryContinue",
-    recovery_continue_task: "notice.decisionRecoveryContinueTask",
-    recovery_revise: "notice.decisionRecoveryRevise",
-    answered: "notice.decisionAnswered",
-  };
-  const outcome = outcomeKeys[receipt.outcome]
-    ? t(outcomeKeys[receipt.outcome] as never)
-    : receipt.outcome || t("notice.decisionReceiptTitle");
-  const showOutcome = receipt.kind !== "ask" || receipt.outcome !== "answered";
-  return (
-    <div className="notice-line__decision-receipt">
-      <span className="notice-line__decision-title">{t(titleKey as never)}</span>
-      {showOutcome && <span className="notice-line__decision-outcome">{outcome}</span>}
-      {receipt.tool && <code>{receipt.tool}</code>}
-      {receipt.subject && <span className="notice-line__decision-subject">{receipt.subject}</span>}
-    </div>
-  );
-}
-
-export function NoticeCard({ item, onAction, actionDisabled = false }: { item: NoticeItem; onAction?: () => void; actionDisabled?: boolean }) {
-  const t = useT();
-  const StatusIcon = item.level === "warn" ? TriangleAlert : Info;
-  const ActionIcon = item.action === "open_changes" ? FileSearch : CirclePlay;
-  return (
-    <div className={`notice-line notice-line--${item.level}${item.variant ? ` notice-line--${item.variant}` : ""}`} data-entrance={item.id}>
-      <StatusIcon className="notice-line__icon" size={14} aria-hidden="true" />
-      <div className="notice-line__text">
-        {item.decisionReceipt ? (
-          <DecisionReceiptLine receipt={item.decisionReceipt} />
-        ) : (
-          <>
-            {item.title ? <div className="notice-line__title">{item.title}</div> : null}
-            <div className="notice-line__body">{item.text}</div>
-          </>
-        )}
-        {item.action && onAction ? (
-          <div className="notice-line__actions">
-            <button className="btn btn--small" type="button" onClick={onAction} disabled={actionDisabled}>
-              <ActionIcon size={13} aria-hidden="true" />
-              <span>{item.action === "open_changes" ? t("notice.completionViewChanges") : t("notice.deliveryIncompleteContinue")}</span>
-            </button>
-          </div>
-        ) : null}
-        {item.detail ? (
-          <details className="notice-line__details">
-            <summary>{t("notice.details")}</summary>
-            <div>{item.detail}</div>
-          </details>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-function CompactionCard({ item }: { item: CompactionItem }) {
-  const t = useT();
-  const [open, setOpen] = useState(false);
-  const beginUserResize = useTranscriptUserResizeIntent();
-  if (item.pending) {
-    return <div className="compaction compaction--pending" data-entrance={item.id}><ProcessCompactIcon size={12} /><span>{t("compaction.working")}</span></div>;
-  }
-  return (
-    <div className="compaction" data-entrance={item.id}>
-      <button type="button" className="compaction__head" onClick={() => { beginUserResize(); setOpen((v) => !v); }} aria-expanded={open}>
-        <ProcessCompactIcon size={12} />
-        <span>{t("compaction.title")}</span>
-        <span className="compaction__meta">{t("compaction.messages", { n: item.messages })}{item.trigger ? ` · ${item.trigger}` : ""}</span>
-        <ChevronRight className={open ? "compaction__chevron--open" : ""} size={12} />
-      </button>
-      {open && <pre className="compaction__body">{item.summary}</pre>}
-    </div>
   );
 }
